# perf(BaseDocument): ~50% faster `as_dict`