## Small, effective changes
- Use `getattr` instead of `doc.get`. Since the columns are valid, there is no reason to assume that `getattr` will be unsafe.
- Define and use new set constants in `frappe.model`
- Remove unreachable code for `is_virtual` fields
- Pass `no_nulls` from `as_dict` to `get_valid_dict`
- Move code which throws error first

## Screenshots

### Before

```python
In [1]: doc = frappe.get_doc("Sales Invoice", "SINV-21-00070")

In [2]: %timeit doc.as_dict()
593 µs ± 700 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)

```

### After

```python
In [1]: doc = frappe.get_doc("Sales Invoice", "SINV-21-00070")

In [2]: %timeit doc.as_dict()
298 µs ± 952 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)

```


[Tested with ERPNext](https://github.com/frappe/erpnext/actions/runs/2115483428)